feat: Rename ipfs to ipmfs to better reflect its naming
