+++ Cargo.toml
@@ -91,11 +91,11 @@ radix_trie = { version = "0.2", optional = true }
 reqsign = "0.4.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+suppaftp = { version = "4.4", features = ["secure"], optional = true }
 thiserror = "1.0"
 time = "0.3"
 tokio = { version = "1.20", features = ["fs"] }
 tracing = { version = "0.1", optional = true }
-suppaftp = {version = "4.4", features = ["secure"], optional = true}
 
 [dev-dependencies]
 cfg-if = "1.0"
