+++ src/operator.rs
@@ -133,7 +133,7 @@ impl Operator {
             Scheme::Http => services::http::Backend::from_iter(it)?.into(),
             #[cfg(feature = "services-ftp")]
             Scheme::Ftp => services::ftp::Backend::from_iter(it)?.into(),
-            Scheme::Ipfs => services::ipfs::Backend::from_iter(it)?.into(),
+            Scheme::Ipmfs => services::ipmfs::Backend::from_iter(it)?.into(),
             Scheme::Memory => services::memory::Builder::default().build()?.into(),
             Scheme::Gcs => services::gcs::Backend::from_iter(it)?.into(),
             Scheme::S3 => services::s3::Backend::from_iter(it)?.into(),
