+++ src/services/fs/backend.rs
@@ -41,8 +41,9 @@ use crate::ops::OpRead;
 use crate::ops::OpStat;
 use crate::ops::OpWrite;
 use crate::ops::Operation;
+use crate::path::build_rel_path;
+use crate::path::build_rooted_abs_path;
 use crate::path::normalize_root;
-use crate::path::{build_rel_path, build_rooted_abs_path};
 use crate::Accessor;
 use crate::BlockingBytesReader;
 use crate::BytesReader;
