+++ src/services/ftp/dir_stream.rs
@@ -37,7 +37,7 @@ pub struct ReadDir {
     files: Vec<String>,
     index: usize,
 }
-/* */
+/*  */
 impl Iterator for ReadDir {
     type Item = Result<File>;
